gcp_cloud_function_for_model_call/function_source/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-aiplatform==1.4.0